refactor/improvements: improvements in binding integration
- Nativehandle dispose calls GetAwaiter().GetResult()
- Updated tests when handles are returned by wrapping it in a Using statement
- Updated Sealed Encryption test's using statement to Item2
- Created Shared Test module and add as reference
- SafeAppPtr as a wrapper with implicit convertor of AppPtr
- Update the URI for Auth test
- In Session.cs implement IDisposible After invoking AppFree set _appPtr to IntPtr.zero- Update tests
-EventHandler must be a static reference in the Session class. IsDisconnected and Disconneted EventHandler should be exposed public

+++ SafeApp.MockAuthBindings/MockAuthBindings.cs
@@ -1,8 +1,9 @@
+using SafeApp.Utilities;
 using System;
 using System.Collections.Generic;
+using System.Linq;
 using System.Runtime.InteropServices;
 using System.Threading.Tasks;
-using SafeApp.Utilities;
 
 namespace SafeApp.MockAuthBindings {
   public partial class MockAuthBindings : IMockAuthBindings {
@@ -16,7 +17,7 @@ namespace SafeApp.MockAuthBindings {
     internal static extern int TestCreateAppNative(out IntPtr oApp);
 
     [DllImport(DllName, EntryPoint = "test_create_app_with_access")]
-    internal static extern int TestCreateAppWithAccessNative([MarshalAs(UnmanagedType.LPArray, SizeParamIndex = 1)] ContainerPermissions[] accessInfo, IntPtr accessInfoLen, out IntPtr oApp);
+    internal static extern int TestCreateAppWithAccessNative([MarshalAs(UnmanagedType.LPArray, SizeParamIndex = 1)] ContainerPermissions[] accessInfo, ulong accessInfoLen, out IntPtr oApp);
 
   }
 }
